base rule for passrole
